Add files via upload
Program for connecting ESP32 to a known local AP and to host a web server with switches to control GPIO of the ESP. IN the current example 2 LEDs connected to pins 26 & 27 are toggled using the webpage button. This can be used to basically control any GPIO functions of the ESP32. The webpage can be accesses through any device on the same local network. Type in ESP IP in device browser